Fix configure for symlinked /usr/bin64/
On some systems /usr/bin64 (which is a fallback search path
for liibqwt.so on configure) is a symlink. "find" exits
directly when passed a symlink without trailing slash, while
a path with trailing slash will resolve the symlink and
search the directory correctly.

This does not affect systems with a regular /usr/lib64.